Add label to Launch screen, move storage helper logic to appropriate class
